move class to HealthEndpoint
