+++ management/src/test/groovy/io/micronaut/management/health/indicator/service/ServiceReadyHealthIndicatorSpec.groovy
@@ -1,6 +1,7 @@
 package io.micronaut.management.health.indicator.service
 
 import io.micronaut.context.ApplicationContext
+import io.micronaut.management.endpoint.health.HealthEndpoint
 import spock.lang.Specification
 
 class ServiceReadyHealthIndicatorSpec extends Specification {
@@ -9,8 +10,20 @@ class ServiceReadyHealthIndicatorSpec extends Specification {
         ApplicationContext applicationContext = ApplicationContext.run(['endpoints.health.service.enabled': 'false'])
 
         expect:
-        !applicationContext.containsBean(ServiceReadyHealthIndicatorConfiguration)
+        !applicationContext.containsBean(HealthEndpoint.ServiceReadyHealthIndicatorConfiguration)
         !applicationContext.containsBean(ServiceReadyHealthIndicator)
+
+        cleanup:
+        applicationContext.close()
+    }
+
+    void "bean of type ServiceReadyHealthIndicatorConfiguration exists by default"() {
+        given:
+        ApplicationContext applicationContext = ApplicationContext.run()
+
+        expect:
+        applicationContext.containsBean(HealthEndpoint.ServiceReadyHealthIndicatorConfiguration)
+        applicationContext.containsBean(ServiceReadyHealthIndicator)
 
         cleanup:
         applicationContext.close()
